Don't install from github when manage_repo is false
I think the general user expectation is that setting `manage_repo =>
false` will allow them to configure their own repo, or bind to a pulp or
spacewalk channel etc somewhere in their profile class.

With this change, `package_source` now defaults to `undef` for all
Operating Systems.  If the user sets a `package_source` that will always
be used and no repo configured.