Support `restart == true` with ProxySQL 2
When using packages that use systemd, and `restart == true`, a systemd
drop-in file is created to add the `--reload` flag to the start command.
Previously, the module assumed init.d only and the `start` parameter of
the Service resource started proxysql manually.  This resulted in the
Pidfile not being written and the systemd service reporting an incorrect
status.

camptocamp/systemd is a new dependency required on modern OSes with
latest ProxySQL.

Fixes #115